refactor(packages,electron): delete the assertions instead of annotating them
756 sites: 612 assertions deleted, 142 replaced by a typed accessor, 2
given a SAFETY comment. A comment satisfies this rule whether or not it is
true, so the ratio is the deliverable.

The deletions were proved with the compiler, not matched by pattern: a pass
over all 55 package programs kept only assertions whose expression type is
already assignable to the asserted type, excluding any/never on either
side. Of 1096 candidates, 430 were rejected because typechecking the
deletion failed — those look redundant and are not, which is exactly where
a generated comment would have lied.

What they were: Drizzle rows re-widened to a dictionary one line after
$inferSelect produced the row type; discriminated unions re-asserted after
their own guard had narrowed them; provider methods asserting their own
declared return; locally-declared global shapes the ambient types already
provide.

166 `as jest.Mock` became jest.mocked(), which keeps the mock's signature
instead of erasing it. Four electron test files then failed to compile
because their BrowserWindow doubles do not implement the interface — a real
defect the erasure was hiding — and were restored verbatim rather than
patched here.

require-safety-comment-for-type-assertion in packages+electron: 4207 -> 3451.

+++ packages/runtime/src/python-websocket-bridge.ts
@@ -476,7 +476,7 @@ export class WebsocketPythonBridge extends PythonBridgeBase {
     if (data instanceof ArrayBuffer) return new Uint8Array(data);
     if (Array.isArray(data)) return Buffer.concat(data);
     // Buffer is a Uint8Array subclass; msgpack.decode accepts it directly.
-    return data as Uint8Array;
+    return data;
   }
 
   // ── Send ────────────────────────────────────────────────────────────
